# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarketCreate": {
        "properties": {
          "game_id": {
            "exclusiveMinimum": 0.0,
            "title": "Game Id",
            "type": "integer"
          },
          "market_type": {
            "maxLength": 48,
            "minLength": 1,
            "title": "Market Type",
            "type": "string"
          },
          "period": {
            "default": "game",
            "maxLength": 24,
            "minLength": 1,
            "title": "Period",
            "type": "string"
          },
          "player_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Id"
          },
          "status": {
            "$ref": "#/components/schemas/MarketStatus",
            "default": "open"
          },
          "team_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          }
        },
        "required": [
          "game_id",
          "market_type"
        ],
        "title": "MarketCreate",
        "type": "object"
      },
      "MarketResponse": {
        "properties": {
          "game_id": {
            "title": "Game Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "market_type": {
            "title": "Market Type",
            "type": "string"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "player_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Id"
          },
          "selections": {
            "items": {
              "$ref": "#/components/schemas/SelectionResponse"
            },
            "title": "Selections",
            "type": "array"
          },
          "status": {
            "$ref": "#/components/schemas/MarketStatus"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          }
        },
        "required": [
          "id",
          "game_id",
          "market_type",
          "period",
          "status"
        ],
        "title": "MarketResponse",
        "type": "object"
      },
      "MarketStatus": {
        "enum": [
          "open",
          "suspended",
          "closed"
        ],
        "title": "MarketStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "DGN-PICKS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/games/{game_id}/markets": {
      "get": {
        "operationId": "list_game_markets_api_v1_games__game_id__markets_get",
        "parameters": [
          {
            "in": "path",
            "name": "game_id",
            "required": true,
            "schema": {
              "title": "Game Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MarketResponse"
                  },
                  "title": "Response List Game Markets Api V1 Games  Game Id  Markets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Game Markets",
        "tags": [
          "games"
        ]
      }
    },
    "/api/v1/markets": {
      "get": {
        "operationId": "list_markets_api_v1_markets_get",
        "parameters": [
          {
            "in": "query",
            "name": "game_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Game Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MarketResponse"
                  },
                  "title": "Response List Markets Api V1 Markets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Markets",
        "tags": [
          "markets"
        ]
      },
      "post": {
        "operationId": "create_market_api_v1_markets_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-dgn-write-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dgn-Write-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Market",
        "tags": [
          "markets"
        ]
      }
    },
    "/api/v1/markets/{market_id}": {
      "delete": {
        "operationId": "delete_market_api_v1_markets__market_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "market_id",
            "required": true,
            "schema": {
              "title": "Market Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-dgn-write-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dgn-Write-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Market",
        "tags": [
          "markets"
        ]
      },
      "get": {
        "operationId": "get_market_api_v1_markets__market_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "market_id",
            "required": true,
            "schema": {
              "title": "Market Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Market",
        "tags": [
          "markets"
        ]
      },
      "patch": {
        "operationId": "update_market_api_v1_markets__market_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "market_id",
            "required": true,
            "schema": {
              "title": "Market Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-dgn-write-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dgn-Write-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Market",
        "tags": [
          "markets"
        ]
      }
    },
    "/api/v1/markets/{market_id}/history": {
      "get": {
        "operationId": "market_history_api_v1_markets__market_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "market_id",
            "required": true,
            "schema": {
              "title": "Market Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OddsSnapshotResponse"
                  },
                  "title": "Response Market History Api V1 Markets  Market Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market History",
        "tags": [
          "markets"
        ]
      },
      "post": {
        "operationId": "append_market_snapshot_api_v1_markets__market_id__history_post",
        "parameters": [
          {
            "in": "path",
            "name": "market_id",
            "required": true,
            "schema": {
              "title": "Market Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-dgn-write-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dgn-Write-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OddsSnapshotCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OddsSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Market Snapshot",
        "tags": [
          "markets"
        ]
      }
    },
    "/api/v1/markets/{market_id}/selections": {
      "post": {
        "operationId": "create_selection_api_v1_markets__market_id__selections_post",
        "parameters": [
          {
            "in": "path",
            "name": "market_id",
            "required": true,
            "schema": {
              "title": "Market Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-dgn-write-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dgn-Write-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Selection",
        "tags": [
          "markets"
        ]
      }
    },
...
    "/api/v1/selections": {
      "get": {
        "operationId": "list_selections_api_v1_selections_get",
        "parameters": [
          {
            "in": "query",
            "name": "market_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SelectionResponse"
                  },
                  "title": "Response List Selections Api V1 Selections Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Selections",
        "tags": [
          "selections"
        ]
      }
    },
    "/api/v1/selections/{selection_id}": {
      "delete": {
        "operationId": "delete_selection_api_v1_selections__selection_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "selection_id",
            "required": true,
            "schema": {
              "title": "Selection Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-dgn-write-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dgn-Write-Key"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Selection",
        "tags": [
          "selections"
        ]
      },
      "get": {
        "operationId": "get_selection_api_v1_selections__selection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "selection_id",
            "required": true,
            "schema": {
              "title": "Selection Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Selection",
        "tags": [
          "selections"
        ]
      },
      "patch": {
        "operationId": "update_selection_api_v1_selections__selection_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "selection_id",
            "required": true,
            "schema": {
              "title": "Selection Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-dgn-write-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dgn-Write-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Selection",
        "tags": [
          "selections"
        ]
      }
    },
...
  }
}